# xtasks-openapi.yaml
openapi: 3.0.0
info:
  title: task-management
  version: '1.0'
  contact:
    email: 34454174+grafma8@users.noreply.github.com
  description: sample api for task-management
servers:
  - url: 'http://localhost:3000'
...
paths:
  '/users/{user_id}':
    get:
      summary: get_user_info
      tags:
        - user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                title: User
                type: object
                x-examples:
                  example-1:
                    user_id: 1
                    user_name: string
                    email: string
                    user_type: 0
                    user_status: 0
                properties:
                  user_id:
                    type: number
                  user_name:
                    type: string
                  email:
                    type: string
                  user_type:
                    type: number
                  user_status:
                    type: number
              examples:
                example-1:
                  value:
                    user_id: 0
                    user_name: string
                    email: string
                    user_type: 0
                    user_status: 0
            application/xml:
              schema:
                type: object
                properties: {}
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      operationId: get-users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples: {}
      description: get user information with user_id
    parameters:
      - schema:
          type: string
        name: user_id
        in: path
        required: true
...
components:
...
  securitySchemes: {}